# internal/infrastructure/api/openapi/api.yaml
openapi: 3.0.3
info:
  title: UrlShortener
  description: OpenAPI Url shortener
  version: 1.0.0
tags:
  - name: ShortURL
    description: Creating short URL
  - name: Stats
    description: Getting statistics about URL

paths:
  /:
    post:
      summary: Create short URL
      tags: 
        - ShortURL
      operationId: CreateshortURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestURL'
      responses:
        '200':
          description: short url created
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/ResponseURL'
        '400':
          description: bad request
          content: {}
        '500':
          description: Internal server error
          content: {}
  /{short-url}:
    get:
      summary: Redirect to original URL by short URL
      tags:
        - ShortURL
      operationId: RedirectURL
      parameters:
        - name: short-url
          in: path
          description: Short URL for redurect
          required: true
          schema:
            type: string
      responses:
        303:
          description: successful operation
        404:
          description: not found
        500:
          description: internal server error
  /stats/{short-url}:
    get:
      summary: Get url statistics
      tags:
        - Stats
      operationId: GetStats
      parameters:
        - name: short-url
          in: path
          description: url for getting stats
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/Stats'
        404:
          description: not found
        500:
          description: internal server error

          
components:
  schemas:
    RequestURL:
      type: object
      properties:
        originalURL:
          type: string
          format: url
    ResponseURL:
      type: object
      properties:
        shortURL:
          type: string
          format: url
    Stats:
      type: object
      properties:
        shortURL:
          type: string
          format: url
        originalURL:
          type: string
          format: url
        statistics: 
          type: integer
          format: int64